Refactor mlesolver into package
