+++ mlesolver/solver.py
@@ -1,204 +1,11 @@
 import random
 from copy import copy
-from copy import deepcopy
 from common_imports import *
-from abc import abstractmethod
-
-
 from tools.common import execute_code
-from inference import *
-from pathlib import Path
-
-
-from contextlib import contextmanager
-import sys, os
-
-
-os.environ["JOBLIB_VERBOSITY"] = "0"
-logging.basicConfig(level=logging.WARNING)
-warnings.filterwarnings("ignore")
-warnings.simplefilter(action='ignore', category=FutureWarning)
-import logging
-logging.getLogger('sklearn.model_selection').setLevel(logging.WARNING)
-
-
-GLOBAL_REPAIR_ATTEMPTS = 2
-
-
-class Command:
-    def __init__(self):
-        self.cmd_type = "OTHER"
-
-    @abstractmethod
-    def docstring(self) -> str:
-        pass
-
-    @abstractmethod
-    def execute_command(self, *args) -> str:
-        pass
-
-    @abstractmethod
-    def matches_command(self, cmd_str) -> bool:
-        pass
-
-    @abstractmethod
-    def parse_command(self, cmd_str) -> tuple:
-        pass
-
-
-"""
-@@@@@@@@@@@@@@@@@@
-@@ CODING TOOLS @@
-@@@@@@@@@@@@@@@@@@
-"""
-
-class Replace(Command):
-    def __init__(self):
-        super().__init__()
-        self.cmd_type = "CODE-replace"
-
-    def docstring(self) -> str:
-        return (
-            "============= REWRITE CODE EDITING TOOL =============\n"
-            "You also have access to a code replacing tool. \n"
-            "This tool allows you to entirely re-write/replace all of the current code and erase all existing code.\n"
-            "You can use this tool via the following command: ```REPLACE\n<code here>\n```, where REPLACE is the word REPLACE and <code here> will be the new code that is replacing the entire set of old code. This tool is useful if you want to make very significant changes, such as entirely changing the model, or the learning process. Before changing the existing code to be your new code, your new code will be tested and if it returns an error it will not replace the existing code. Try limiting the use of rewriting and aim for editing the code more."
-        )
-
-    def execute_command(self, *args) -> str:
-        # args[0] -> new code
-        args = args[0]
-        return args[0]
-
-    def matches_command(self, cmd_str) -> bool:
-        if "```REPLACE" in cmd_str: return True
-        return False
-
-    def parse_command(self, *args) -> tuple:
-        new_code = extract_prompt(args[0], "REPLACE")
-        code_exec = f"{args[1]}\n{new_code}"
-        code_ret = execute_code(code_exec)
-        if "[CODE EXECUTION ERROR]" in code_ret: return False, (None, code_ret,)
-        return True, (new_code.split("\n"), code_ret)
-
-
-
-class Edit(Command):
-    def __init__(self):
-        super().__init__()
-        self.cmd_type = "CODE-edit"
-
-    def docstring(self) -> str:
-        return (
-            "============= CODE EDITING TOOL =============\n"
-            "You also have access to a code editing tool. \n"
-            "This tool allows you to replace lines indexed n through m (n:m) of the current code with as many lines of new code as you want to add. This removal is inclusive meaning that line n and m and everything between n and m is removed. This will be the primary way that you interact with code. \n"
-            "You can edit code using the following command: ```EDIT N M\n<new lines to replace old lines>\n``` EDIT is the word EDIT, N is the first line index you want to replace and M the the last line index you want to replace (everything inbetween will also be removed), and <new lines to replace old lines> will be the new code that is replacing the old code. Before changing the existing code to be your new code, your new code will be tested and if it returns an error it will not replace the existing code. Your changes should significantly change the functionality of the code."
-        )
-
-    def execute_command(self, *args) -> str:
-        # args[0] -> N (int)
-        # args[1] -> M (int)
-        # args[2] -> old code
-        # args[3] -> new lines to replace
-        # args[4] -> new lines to replace
-        try:
-            args = args[0]
-            current_code = args[2]
-            lines_to_add = list(reversed(args[3]))
-            lines_to_replace = list(reversed(range(args[0], args[1]+1)))
-            for _ln in lines_to_replace:
-                current_code.pop(_ln)
-            for _line in lines_to_add:
-                current_code.insert(args[0], _line)
-            new_code = "\n".join(current_code)
-            code_exec = f"{args[4]}\n{new_code}"
-            code_ret = execute_code(code_exec)
-            if "CODE EXECUTION ERROR" in code_ret: return (False, None, code_ret)
-            return (True, current_code, code_ret)
-        except Exception as e:
-            return (False, None, str(e))
-
-    def matches_command(self, cmd_str) -> bool:
-        if "```EDIT" in cmd_str: return True
-        return False
-
-    def parse_command(self, *args) -> tuple:
-        cmd_str, codelines, datasetcode = args[0], args[1], args[2]
-        success = True
-        try:
-            text = extract_prompt(cmd_str, "EDIT").split("\n")
-            if len(text) == 0: return False, None
-            lines_to_edit = text[0].split(" ")
-            if len(lines_to_edit) != 2: return False, None
-            lines_to_edit = [int(_) for _ in lines_to_edit]
-            if len(text[1:]) == 0: return False, None
-            return success, (lines_to_edit[0], lines_to_edit[1], codelines, text[1:], datasetcode)
-        except Exception as e:
-            return False, (None, None, None, None, None)
-
-
-def get_score(outlined_plan, code, code_return, REWARD_MODEL_LLM, attempts=3, openai_api_key=None):
-    e = str()
-    for _attempt in range(attempts):
-        try:
-            # todo: have a reward function here
-            sys = (
-                f"You are a professor agent who is serving as an expert reward model that can read a research plan, research code, and code output and are able to determine how well a model followed the plan, built the code, and got the proper output scored from 0 to 1 as a float.\n\n"
-                f"You must structure your score exactly in the following way: ```SCORE\n<score here>\n``` where SCORE is just the word score, <score here> is a floating point number between 0 and 1 representing how well the model followed the plan, built the code, and got the proper output."
-            )
-            scoring = query_model(
-                model_str=f"{REWARD_MODEL_LLM}",
-                system_prompt=sys,
-                openai_api_key=openai_api_key,
-                prompt=(
-                    f"Outlined in the following text is the research plan that the machine learning engineer was tasked with building: {outlined_plan}\n\n"
-                    f"The following text is the research code that the model produced: \n{code}\n\n"
-                    f"The following is the output from the model: {code_return}\n\n"), temp=0.6)
-            performance = extract_prompt(text=scoring, word="SCORE")
-            performance = float(performance)
-            return performance, f"The performance of your submission is: {performance}", True
-        except Exception as e:
-            return None, str(e), False
-    return 0, e
-
-
-def code_repair(code, error, ctype, REPAIR_LLM, openai_api_key=None):
-    if ctype == "replace":
-        repair_sys = (
-            "You are an automated code repair tool.\n"
-            "Your goal is to take in code and an error and repair the code to make sure the same error does not repeat itself, and also to remove any other potential errors from the code without affecting the code output.\n"
-            "Your output should match the original code as closely as possible.\n"
-            "You must wrap the code in the following ```python\n<code here>\n```\n"
-            "Do not forget the opening ```python and the closing ```."
-        )
-        model_resp = query_model(
-            openai_api_key=openai_api_key,
-            model_str=f"{REPAIR_LLM}",
-            system_prompt=repair_sys,
-            prompt=f"Provided here is the error: {error}\n\nProvided below is the code:\n\n{code}", temp=0.8)
-        return extract_prompt(model_resp, "python")
-    elif ctype == "edit":
-        repair_sys = (
-            "You are an automated code repair tool.\n"
-            "Your goal is to take in code and an error and repair the code to make sure the same error does not repeat itself, and also to remove any other potential errors from the code without affecting the code output.\n"
-            "Your output should match the original code as closely as possible.\n"
-            
-            "============= CODE EDITING TOOL =============\n"
-            "You have access to a code editing tool. \n"
-            "This tool allows you to replace lines indexed n through m (n:m) of the current code with as many lines of new code as you want to add. This removal is inclusive meaning that line n and m and everything between n and m is removed. This will be the primary way that you interact with code. \n"
-            "You can edit code using the following command: ```EDIT N M\n<new lines to replace old lines>\n``` EDIT is the word EDIT, N is the first line index you want to replace and M the the last line index you want to replace (everything inbetween will also be removed), and <new lines to replace old lines> will be the new code that is replacing the old code. Before changing the existing code to be your new code, your new code will be tested and if it returns an error it will not replace the existing code.\n"
-            "Please use the code editing tool to fix this code."
-            "Do not forget the opening ```EDIT N M and the closing ```."
-            "Your output should look like the following\n\n```EDIT N M\n<new lines to replace old lines>\n```"
-        )
-        model_resp = query_model(
-            openai_api_key=openai_api_key,
-            model_str=f"{REPAIR_LLM}",
-            system_prompt=repair_sys,
-            prompt=f"Provided here is the error: {error}\n\nProvided below is the code:\n\n{code}", temp=0.2)
-        return model_resp
-
+from inference import query_model
+from utils import remove_figures, extract_prompt
+from .commands import Replace, Edit
+from .utils import get_score, code_repair, GLOBAL_REPAIR_ATTEMPTS
 
 class MLESolver:
     def __init__(self, dataset_code, openai_api_key=None, notes=None, max_steps=10, insights=None, plan=None, llm_str=None):
